requirements.txt
faster-whisper==1.1.1
llama-cpp-python==0.3.4
openwakeword==0.6.0
sounddevice==0.5.1
soundfile==0.12.1
numpy==1.26.4
onnxruntime==1.20.1
requests==2.32.3